Various optimisation to the code

+++ rank/interfaces.go
@@ -3,7 +3,7 @@ package rank
 // GetRanker provide a rank operation and a get operation over a bitset.
 type GetRanker interface {
 	// Returns the position of the i'th bit in the underlying bitset.
-	Get(i int) int
+	Get(i int) uint64
 	// Returns the number of high bits between [0, position] in the underlying bitset.
 	Rank(position int) int
 }
@@ -11,7 +11,7 @@ type GetRanker interface {
 // GetRankSelecter provide get, rank and select operations over a bitset.
 type GetRankSelecter interface {
 	// Returns the position of the i'th bit in the underlying bitset.
-	Get(i int) int
+	Get(i int) uint64
 	// Returns the number of high bits between [0, position] in the underlying bitset.
 	Rank(position int) int
 	// Returns the position of the i'th high bit in the underlying bitset.
